Refactor with PR Comments     - Moved ContractWrapper into hidden module and renamed it         BaseContractWrapper     - Moved TxParams into its own module     - Make sure docs show public inherited functions and update         them to reflect PR comments
