+++ python-packages/contract_wrappers/test/test_base_contract_wrapper.py
@@ -1,31 +1,33 @@
 """
-Tests for :class:`ContractWrapper`.
+Tests for :class:`BaseContractWrapper`.
 """
 
 import pytest
 from eth_utils import to_checksum_address
 
 from zero_ex.contract_artifacts import abi_by_name
-from zero_ex.contract_wrappers import ContractWrapper
+from zero_ex.contract_wrappers._base_contract_wrapper import (
+    BaseContractWrapper,
+)
 
 
 @pytest.fixture(scope="module")
 def contract_wrapper(ganache_provider):
-    """Get a ContractWrapper instance for testing."""
-    return ContractWrapper(provider=ganache_provider)
+    """Get a BaseContractWrapper instance for testing."""
+    return BaseContractWrapper(provider=ganache_provider)
 
 
 def test_contract_wrapper__get_accounts(
     accounts, contract_wrapper
 ):  # pylint: disable=redefined-outer-name
-    """Test getting list of accounts from ContractWrapper instance."""
+    """Test getting list of accounts from BaseContractWrapper instance."""
     assert contract_wrapper.get_accounts() == accounts
 
 
 def test_contract_wrapper__get_default_account(
     accounts, contract_wrapper
 ):  # pylint: disable=redefined-outer-name
-    """Test getting default account from ContractWrapper instance."""
+    """Test getting default account from BaseContractWrapper instance."""
     assert contract_wrapper.get_default_account() == accounts[0]
 
 
@@ -35,7 +37,7 @@ def test_contract_wrapper__execute_method(
     erc20_proxy_address,
     weth_address,  # pylint: disable=redefined-outer-name
 ):
-    """Test :function:`ContractWrapper.execute` method."""
+    """Test :function:`BaseContractWrapper.execute` method."""
     acc1_allowance = contract_wrapper.execute_method(
         address=weth_address,
         abi=abi_by_name("WETH9"),
